Make selected tools operations appear on the primary classes to which they pertain.
The tools operations are still defined by Scufl2Tools and URITools, but
now the workflow bundle will construct a standard instance of them and
make that available to its children, which can use it to do the relevant
operations. This makes it less necessary for users of the library to
know about how the serialization model of Scufl2 works.

Not all tools operations are exposed this way.