Update JavaDoc in Processor.java - rm @value tags
Changed two statements to remove @value tag
and use just field names. 
The @value tags are not working. As a result the getType()
Javadoc statements are incomplete, rendering as, 
"...the value given by ." and 
"...the type given by the constant ."
Problem may be caused because the fields are
not literal values. 
(Ref: http://stackoverflow.com/questions/19807696/how-to-use-value-tag-in-javadoc)